Implement threaded playback
thread::spawn returns a join handle that is not copyable.  Therefore
saving it, moving it and calling join() on it is not trivial.  Using
Option and std::mem::swap did the trick in the end.

In addition, flooding ALSA with samples is prevented by specifying the
period and the buffer size.