Fix playback and use the given pattern
Write size to ALSA device is now independent of the playback buffer size
which guarantees responsiveness when stopping the playback.  Before
playback was only stopped when the whole buffer was finished being
written to the ALSA device.

Beatplayer's pattern member is now honored when creating the playback
buffer.  A necessary step for implementing the `pattern` command.