Enhance BeatPlayer's pattern and fix Repl's argument parsing
The pattern in BeatPlayer now support pause as well.  The
implementation consists of a custom type BeatPattern and enum
BeatPatternType.  Idiomatic conversion from str slice to BeatPattern is
done via TryFrom trait.

The Repl's argument parsing is now correct.  It was not buggy e.g.
because it removed whitespaces within the argument because the
str::split function usage.