Current Beat is Marked In Status Line
The current beat is marked via an empty space on each side in the status
line in the pattern segment.

The implementation uses std::time::Instant as to generate the elapsed
time and calculates the beat currently played beat index to be marked in
the status line.  This approach might be error prone on platform where
std::time::Instant is not a monotonic clock.  A more resilient approach
might be needed.